feat: add reply function to post
