+++ app/(home)/page.tsx
@@ -4,6 +4,7 @@ import PostForm from '@/src/compositions/post/post-form';
 import { Heading, HeadingLevel } from '@ost-cas-fee-adv-23-24/design-system-pixelpioneers';
 import { getUser } from '../actions/user';
 import { MessageVariant } from '@/src/compositions/post/types';
+import { createPost } from '@/app/actions/post';
 
 export default async function Home() {
     const session = await auth();
@@ -23,7 +24,11 @@ export default async function Home() {
             </header>
             {user && (
                 <section className="flex w-full flex-col gap-y-m px-m md:w-auto md:px-0">
-                    <PostForm user={user} messageVariant={MessageVariant.CREATE} />
+                    <PostForm
+                        user={user}
+                        messageVariant={MessageVariant.POST}
+                        onCreate={createPost}
+                    />
                 </section>
             )}
         </>
